Tidy up frontend
Split login and registration into separate routes. This will make the
code more readable.
