+++ client/src/components/login/login.js
@@ -1,0 +1,53 @@
+import React, { useState } from "react";
+
+export default function Login() {
+  const [loginUsername, setLoginUsername] = useState("");
+  const [loginPassword, setLoginPassword] = useState("");
+  const [data, setData] = useState(null);
+
+  const login = () => {
+    fetch("http://localhost:9000/login", {
+      method: "POST",
+      credentials: "include",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({
+        username: loginUsername,
+        password: loginPassword,
+      }),
+    }).then((res) => console.log(res));
+  };
+
+  const getUser = () => {
+    fetch("http://localhost:9000/user", {
+      method: "GET",
+      credentials: "include",
+      headers: { "Content-Type": "application/json" },
+    })
+      .then((res) => {
+        return res.json();
+      })
+      .then((data) => setData(data));
+  };
+
+  return (
+    <div>
+      <div>
+        <h1>Login</h1>
+        <input
+          placeholder="username"
+          onChange={(e) => setLoginUsername(e.target.value)}
+        ></input>
+        <input
+          placeholder="password"
+          onChange={(e) => setLoginPassword(e.target.value)}
+        ></input>
+        <button onClick={login}>Submit</button>
+      </div>
+      <div>
+        <h1>Get User</h1>
+        <button onClick={getUser}>Submit</button>
+        {data ? <h1>Welcome Back {data.username}</h1> : null}
+      </div>
+    </div>
+  );
+}
